Fix return type of length operations to be Py_ssize_t.
git-svn-id: http://svn.code.sf.net/p/cxx/code@356 cb2f5747-f50d-0410-b093-b27f9c363ba5

+++ trunk/CXX/CXX/Python3/ExtensionTypeBase.hxx
@@ -89,7 +89,7 @@ namespace Py
         virtual PyObject *iternext();
 
         // Sequence methods
-        virtual int sequence_length();
+        virtual Py_ssize_t sequence_length();
         virtual Object sequence_concat( const Object & );
         virtual Object sequence_repeat( Py_ssize_t );
         virtual Object sequence_item( Py_ssize_t );
@@ -99,7 +99,7 @@ namespace Py
         virtual Object sequence_inplace_repeat( Py_ssize_t );
 
         // Mapping
-        virtual int mapping_length();
+        virtual Py_ssize_t mapping_length();
         virtual Object mapping_subscript( const Object & );
         virtual int mapping_ass_subscript( const Object &, const Object & );
