+++ trunk/CXX/Demo/Python3/range.hxx
@@ -60,7 +60,7 @@ public:
     virtual Py::Object repr();
     virtual Py::Object getattr( const char *name );
 
-    virtual int sequence_length();
+    virtual Py_ssize_t sequence_length();
     virtual Py::Object sequence_item( Py_ssize_t i );
     virtual Py::Object sequence_concat( const Py::Object &j );
     virtual Py::Object sequence_slice( Py_ssize_t i, Py_ssize_t j );
