+++ trunk/CXX/Src/Python2/cxx_extensions.cxx
@@ -1523,7 +1523,7 @@ PyObject *PythonExtensionBase::iternext()
 }
 
 // Sequence methods
-int PythonExtensionBase::sequence_length()
+Py_ssize_t PythonExtensionBase::sequence_length()
 {
     missing_method( sequence_length );
 }
@@ -1569,7 +1569,7 @@ Object PythonExtensionBase::sequence_inplace_repeat( Py_ssize_t )
 }
 
 // Mapping
-int PythonExtensionBase::mapping_length()
+Py_ssize_t PythonExtensionBase::mapping_length()
 {
     missing_method( mapping_length );
 }
